Release 1.2.4
+ The commands required to commit/save the configuration on a device
are now attached to NetDevice objects under the commit_commands
attribute, to make it easier to execute these commands without having
to determine them for yourself.
+ Added a way to optionally perform a "commit full" operation on
Juniper devices by defining a dictionary of attributes and values for
matching devices using settings.JUNIPER_FULL_COMMIT_FIELDS. This
modifies the commit_commands that are assigned when the NetDevice
object is created.
+ Console paging is now disabled by default for async SSH channels.